+++ packages/core/src/index.ts
@@ -6,6 +6,7 @@ export { combineRoutes } from './router/router.combiner';
 export { combineEffects, combineMiddlewares } from './effects/effects.combiner';
 export { createEffectMetadata } from './effects/effectsMetadata.factory';
 export * from './effects/effects.interface';
+export * from './effects/http-effects.interface';
 export * from './router/router.interface';
 export * from './operators';
 export * from './http.interface';
