+++ packages/core/src/router/router.resolver.ts
@@ -1,7 +1,8 @@
 import { EMPTY, Observable, of } from 'rxjs';
 import { mergeMap, takeWhile } from 'rxjs/operators';
 import { HttpMethod, HttpRequest, HttpResponse } from '../http.interface';
-import { HttpEffectResponse, EffectMetadata } from '../effects/effects.interface';
+import { EffectMetadata } from '../effects/effects.interface';
+import { HttpEffectResponse } from '../effects/http-effects.interface';
 import { RouteMatched, Routing, RoutingItem } from './router.interface';
 import { queryParamsFactory } from './router.query.factory';
 export { RoutingItem };
